move db-related code & templates into `oio_rest.db` package
