+++ oio_rest/oio_rest/db/sql/post-funcs/_ensure_document_del_exists_and_get.sql
