+++ oio_rest/oio_rest/db/sql/pre-funcs/_as_search_match_array.sql
