+++ oio_rest/oio_rest/db/sql/pre-funcs/_as_valid_registrering_livscyklus_transition.sql
