+++ oio_rest/oio_rest/db/sql/templates/_as_create_registrering.jinja.sql
