+++ oio_rest/oio_rest/db/sql/templates/_as_get_prev_registrering.jinja.sql
