+++ oio_rest/oio_rest/db/sql/templates/_remove_nulls_in_array.jinja.sql
