+++ oio_rest/oio_rest/db/sql/templates/as_list.jinja.sql
