+++ oio_rest/oio_rest/db/templates/attribute_array.sql
