+++ oio_rest/oio_rest/db/templates/create_object.sql
