+++ oio_rest/oio_rest/db/templates/list_objects.sql
