+++ oio_rest/oio_rest/db/templates/registration.sql
