+++ oio_rest/oio_rest/db/templates/search_objects.sql
