+++ oio_rest/oio_rest/db/templates/state_array.sql
