+++ oio_rest/oio_rest/utils/build_registration.py
@@ -10,11 +10,11 @@ import itertools
 import uuid
 from werkzeug.datastructures import MultiDict
 
-from oio_rest.db_helpers import get_attribute_names, get_attribute_fields
-from oio_rest.db_helpers import get_state_names, get_relation_names
-from oio_rest.db_helpers import get_document_part_relation_names
-from oio_rest.db_helpers import DokumentVariantEgenskaberType
-from oio_rest.db_helpers import DokumentDelEgenskaberType
+from oio_rest.db.db_helpers import get_attribute_names, get_attribute_fields
+from oio_rest.db.db_helpers import get_state_names, get_relation_names
+from oio_rest.db.db_helpers import get_document_part_relation_names
+from oio_rest.db.db_helpers import DokumentVariantEgenskaberType
+from oio_rest.db.db_helpers import DokumentDelEgenskaberType
 
 
 def is_urn(s):
