+++ oio_rest/tests/test_db.py
@@ -35,7 +35,7 @@ class TestDB(flask_testing.TestCase):
 
     @patch('oio_rest.db.get_relation_field_type')
     def test_convert_relation_value_journalnotat(self, mock_get_rel):
-        from oio_rest.db_helpers import JournalNotat
+        from oio_rest.db.db_helpers import JournalNotat
 
         # Arrange
         mock_get_rel.return_value = "journalnotat"
@@ -55,7 +55,7 @@ class TestDB(flask_testing.TestCase):
 
     @patch('oio_rest.db.get_relation_field_type')
     def test_convert_relation_value_journaldokument(self, mock_get_rel):
-        from oio_rest.db_helpers import JournalDokument, OffentlighedUndtaget
+        from oio_rest.db.db_helpers import JournalDokument, OffentlighedUndtaget
 
         # Arrange
         mock_get_rel.return_value = "journaldokument"
@@ -84,7 +84,7 @@ class TestDB(flask_testing.TestCase):
 
     @patch('oio_rest.db.get_relation_field_type')
     def test_convert_relation_value_aktoerattr(self, mock_get_rel):
-        from oio_rest.db_helpers import AktoerAttr
+        from oio_rest.db.db_helpers import AktoerAttr
 
         # Arrange
         mock_get_rel.return_value = "aktoerattr"
@@ -127,7 +127,7 @@ class TestDB(flask_testing.TestCase):
 
     @patch('oio_rest.db.get_relation_field_type')
     def test_convert_relation_value_vaerdirelationattr(self, mock_get_rel):
-        from oio_rest.db_helpers import VaerdiRelationAttr
+        from oio_rest.db.db_helpers import VaerdiRelationAttr
 
         # Arrange
         mock_get_rel.return_value = "vaerdirelationattr"
@@ -405,7 +405,7 @@ class TestDB(flask_testing.TestCase):
     def test_convert_attr_value_soegeord(self,
                                          mock_get_field_type):
         # type: (MagicMock) -> None
-        from oio_rest.db_helpers import Soegeord
+        from oio_rest.db.db_helpers import Soegeord
 
         # Arrange
         mock_get_field_type.return_value = "soegeord"
@@ -428,7 +428,7 @@ class TestDB(flask_testing.TestCase):
     def test_convert_attr_value_offentlighedundtagettype(self,
                                                          mock_get_field_type):
         # type: (MagicMock) -> None
-        from oio_rest.db_helpers import OffentlighedUndtaget
+        from oio_rest.db.db_helpers import OffentlighedUndtaget
 
         # Arrange
         mock_get_field_type.return_value = "offentlighedundtagettype"
