+++ oio_rest/tests/test_db_helpers.py
@@ -11,9 +11,9 @@ from unittest import TestCase
 from mock import MagicMock, call, patch
 from werkzeug.datastructures import ImmutableMultiDict
 
-from oio_rest import db_helpers
 from oio_rest import settings
 from oio_rest.custom_exceptions import BadRequestException
+from oio_rest.db import db_helpers
 from oio_rest.utils import test_support
 
 
@@ -840,7 +840,7 @@ class TestDBHelpers(TestCase):
             db_helpers.to_bool("This is not a valid boolean value")
 
     def test_dokumentvarianttype_input_when_none(self):
-        from oio_rest.db_helpers import DokumentVariantType
+        from oio_rest.db.db_helpers import DokumentVariantType
 
         # Arrange
         expected_result = None
@@ -852,7 +852,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(expected_result, actual_result)
 
     def test_dokumentvariantegenskabertype_input_when_none(self):
-        from oio_rest.db_helpers import DokumentVariantEgenskaberType
+        from oio_rest.db.db_helpers import DokumentVariantEgenskaberType
 
         # Arrange
         expected_result = None
@@ -864,7 +864,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(expected_result, actual_result)
 
     def test_dokumentdeltype_input_when_none(self):
-        from oio_rest.db_helpers import DokumentDelType
+        from oio_rest.db.db_helpers import DokumentDelType
 
         # Arrange
         expected_result = None
@@ -876,7 +876,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(expected_result, actual_result)
 
     def test_virkning_input_when_none(self):
-        from oio_rest.db_helpers import Virkning
+        from oio_rest.db.db_helpers import Virkning
 
         # Arrange
         expected_result = None
@@ -888,7 +888,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(expected_result, actual_result)
 
     def test_dokumentdelegenskabertype_get_file_storage_raises_bre(self):
-        from oio_rest.db_helpers import DokumentDelEgenskaberType
+        from oio_rest.db.db_helpers import DokumentDelEgenskaberType
         import flask
 
         # Arrange
@@ -901,7 +901,7 @@ class TestDBHelpers(TestCase):
                 'field:not_in_request')
 
     def test_dokumentdelegenskabertype_get_file_storage(self):
-        from oio_rest.db_helpers import DokumentDelEgenskaberType
+        from oio_rest.db.db_helpers import DokumentDelEgenskaberType
         import flask
         from flask import request
 
@@ -920,7 +920,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(mockfile, actual_result)
 
     def test_dokumentdelegenskabertype_get_file_storage_returns_none(self):
-        from oio_rest.db_helpers import DokumentDelEgenskaberType
+        from oio_rest.db.db_helpers import DokumentDelEgenskaberType
 
         # Arrange
 
@@ -931,13 +931,13 @@ class TestDBHelpers(TestCase):
         # Assert
         self.assertIsNone(actual_result)
 
-    @patch('oio_rest.db_helpers.DokumentDelEgenskaberType'
+    @patch('oio_rest.db.db_helpers.DokumentDelEgenskaberType'
            '._get_file_storage_for_content_url')
-    @patch('oio_rest.db_helpers.content_store.save_file_object')
+    @patch('oio_rest.db.db_helpers.content_store.save_file_object')
     def test_dokumentdelegenskabertype_input_update_file(self, mock_save_file,
                                                          mock_get_file):
         # type: (MagicMock, MagicMock) -> None
-        from oio_rest.db_helpers import DokumentDelEgenskaberType
+        from oio_rest.db.db_helpers import DokumentDelEgenskaberType
         import flask
 
         # Arrange
@@ -958,7 +958,7 @@ class TestDBHelpers(TestCase):
         mock_save_file.assert_called_with(mockfile)
 
     def test_dokumentdelegenskabertype_input_when_none(self):
-        from oio_rest.db_helpers import DokumentDelEgenskaberType
+        from oio_rest.db.db_helpers import DokumentDelEgenskaberType
 
         # Arrange
         expected_result = None
@@ -970,7 +970,7 @@ class TestDBHelpers(TestCase):
         self.assertEqual(expected_result, actual_result)
 
     def test_dokumentdelrelationtype_input_when_none(self):
-        from oio_rest.db_helpers import DokumentDelRelationType
+        from oio_rest.db.db_helpers import DokumentDelRelationType
 
         # Arrange
         expected_result = None
@@ -983,12 +983,12 @@ class TestDBHelpers(TestCase):
 
 
 class TestNamedTupleAdapter(TestCase):
-    from oio_rest.db_helpers import NamedTupleAdapter
+    from oio_rest.db.db_helpers import NamedTupleAdapter
 
     class TestClass(NamedTupleAdapter):
         pass
 
-    @patch('oio_rest.db_helpers.psyco_adapt')
+    @patch('oio_rest.db.db_helpers.psyco_adapt')
     def test_prepare_and_adapt(self, mock_psyco_adapt):
         # type: (MagicMock) -> None
         # Arrange
@@ -1027,7 +1027,7 @@ class TestNamedTupleAdapter(TestCase):
 
 
 class TestAktoerAttrAdapter(TestCase):
-    from oio_rest.db_helpers import AktoerAttrAdapter
+    from oio_rest.db.db_helpers import AktoerAttrAdapter
 
     class TestClass(AktoerAttrAdapter):
         pass
@@ -1059,7 +1059,7 @@ class TestAktoerAttrAdapter(TestCase):
 
 
 class TestSearchable(TestCase):
-    from oio_rest.db_helpers import Searchable
+    from oio_rest.db.db_helpers import Searchable
 
     class TestSearchableClass(Searchable):
         _fields = ()
